+++ lib/qhg_correlator.h
@@ -3,7 +3,7 @@
 
 #include <qhg_types.h>
 
-qhg_correlator qhg_correlator_init(int, qhg_lattice *);
+qhg_correlator qhg_correlator_init(size_t, qhg_lattice *);
 qhg_correlator qhg_correlator_copy(qhg_correlator);
 void qhg_correlator_finalize(qhg_correlator);
