+++ lib/qhg_su3_project.h
@@ -2,6 +2,6 @@
 #define _QHG_SU3_PROJECT_H 1
 #include <complex.h>
 
-void qhg_su3_project(_Complex double *u, int);
+void qhg_su3_project(_Complex double *u, unsigned long int);
 
 #endif /* _QHG_SU3_PROJECT_H */
